requirements_streamlit.txt
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
torch>=2.0.0
av>=10.0.0
flask>=2.0.0
psycopg2-binary>=2.9.0
plotly>=5.0.0
